# contracts/openapi/workflow-engine-service.yaml
openapi: 3.1.0
info:
  title: Workflow Engine Service API
  version: 1.0.0
  description: >
    REST API wrapping the Flowable 7.x workflow engine. Provides endpoints for
    managing BPMN deployments, process definitions, process instances, and human
    tasks. All operations are tenant-scoped via the JWT bearer token.
  contact:
    name: Platform Engineering
  license:
    name: Proprietary
servers:
  - url: http://workflow-engine-service:8080
    description: Internal Kubernetes service endpoint
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/process-instances:
    post:
      tags: [instances]
      operationId: startProcessInstance
      summary: Start a new process instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProcessInstanceRequest'
      responses:
        '201':
          description: Process instance started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInstanceDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Process definition not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'

    get:
      tags: [instances]
      operationId: listProcessInstances
      summary: List process instances
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: processDefinitionKey
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, SUSPENDED, COMPLETED]
      responses:
        '200':
          description: Paged list of process instances
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PagedResponse'
                  - type: object
                    properties:
                      content:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProcessInstanceDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/process-instances/{instanceId}:
    get:
      tags: [instances]
      operationId: getProcessInstance
      summary: Get a process instance by ID
      parameters:
        - $ref: '#/components/parameters/InstanceIdPath'
      responses:
        '200':
          description: Process instance detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInstanceDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [instances]
      operationId: cancelProcessInstance
      summary: Cancel (delete) a process instance
      description: Terminates an active or suspended process instance.
      parameters:
        - $ref: '#/components/parameters/InstanceIdPath'
        - name: reason
          in: query
          required: false
          schema:
            type: string
          description: Human-readable cancellation reason recorded in the audit log
      responses:
        '204':
          description: Process instance cancelled
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Instance is already completed and cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/process-instances/{instanceId}/suspend:
    post:
      tags: [instances]
      operationId: suspendProcessInstance
      summary: Suspend an active process instance
      parameters:
        - $ref: '#/components/parameters/InstanceIdPath'
      responses:
        '200':
          description: Process instance suspended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInstanceDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Instance is not in a suspendable state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/process-instances/{instanceId}/activate:
    post:
      tags: [instances]
      operationId: activateProcessInstance
      summary: Reactivate a suspended process instance
      parameters:
        - $ref: '#/components/parameters/InstanceIdPath'
      responses:
        '200':
          description: Process instance activated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProcessInstanceDto'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Instance is not in a suspended state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Tasks ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT issued by Keycloak. Must include `tenant_id` and standard `sub`
        claims. Pass as `Authorization: Bearer <token>`.
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page index
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
      description: Number of items per page
...
    InstanceIdPath:
      name: instanceId
      in: path
      required: true
      schema:
        type: string
      description: Process instance ID
...
  responses:
    BadRequest:
      description: Request body or query parameters failed validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user lacks permission for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    UnprocessableEntity:
      description: Semantically invalid request (e.g. invalid BPMN)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    PagedResponse:
      type: object
      description: Generic envelope for paginated list responses
      required: [content, totalElements, totalPages, page, size]
      properties:
        content:
          type: array
          items: {}
          description: Page of result items (type depends on endpoint)
        totalElements:
          type: integer
          format: int64
          description: Total number of items across all pages
          example: 143
        totalPages:
          type: integer
          description: Total number of pages for the requested page size
          example: 8
        page:
          type: integer
          description: Zero-based index of the current page
          example: 0
        size:
          type: integer
          description: Requested page size
          example: 20

    # ── Deployment ────────────────────────────────────────────────────────────
...
    ProcessInstanceDto:
      type: object
      required: [id, processDefinitionKey, processDefinitionId, tenantId, startTime, status, startedBy]
      properties:
        id:
          type: string
          example: "pi-99a2b"
        processDefinitionKey:
          type: string
          example: "order-fulfilment"
        processDefinitionId:
          type: string
          example: "order-fulfilment:3:abc123"
        businessKey:
          type: string
          description: Optional external correlation key
          example: "ORD-20251101-0042"
          nullable: true
        tenantId:
          type: string
          example: "acme-corp"
        startTime:
          type: string
          format: date-time
          example: "2025-11-01T15:00:00Z"
        endTime:
          type: string
          format: date-time
          nullable: true
          description: Populated when the instance has completed or been cancelled
          example: null
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, COMPLETED, CANCELLED]
          example: "ACTIVE"
        startedBy:
          type: string
          description: User ID of the principal who started the instance
          example: "user-uuid-1234"

    # ── Task ─────────────────────────────────────────────────────────────────
...
    StartProcessInstanceRequest:
      type: object
      required: [processDefinitionKey]
      properties:
        processDefinitionKey:
          type: string
          example: "order-fulfilment"
        businessKey:
          type: string
          description: External correlation key (must be unique within the tenant)
          example: "ORD-20251101-0042"
        variables:
          type: object
          additionalProperties: true
          description: >
            Initial process variables. Values may be any JSON-serialisable type.
          example:
            orderId: "ORD-20251101-0042"
            totalAmount: 1500.00
            priority: "HIGH"
...
    ErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
          description: When the error occurred (server time, UTC)
          example: "2025-11-01T15:05:00Z"
        status:
          type: integer
          description: HTTP status code
          example: 404
        error:
          type: string
          description: Short HTTP reason phrase
          example: "Not Found"
        message:
          type: string
          description: Human-readable description of the error
          example: "Process instance pi-99a2b not found for tenant acme-corp"
        path:
          type: string
          description: Request URI that produced the error
          example: "/api/v1/process-instances/pi-99a2b"